+++ tests/control.c
@@ -3,11 +3,11 @@
 #include <metalang99/util.h>
 
 // ML99_if {
-ML99_assertEq(ML99_if(v(ML99_true), v(24), v(848)), v(24));
-ML99_assertEq(ML99_if(v(ML99_true), v(1549), v(1678)), v(1549));
+ML99_assertEq(ML99_if(ML99_true, v(24), v(848)), v(24));
+ML99_assertEq(ML99_if(ML99_true, v(1549), v(1678)), v(1549));
 
-ML99_assertEq(ML99_if(v(ML99_false), v(516), v(115)), v(115));
-ML99_assertEq(ML99_if(v(ML99_false), v(10), v(6)), v(6));
+ML99_assertEq(ML99_if(ML99_false, v(516), v(115)), v(115));
+ML99_assertEq(ML99_if(ML99_false, v(10), v(6)), v(6));
 // }
 
 // ML99_overload {